WIP: Access endpoints_decoded as uint16_t again
Fetching from this array takes some time.

I figured we could store it as uint16 so that we only need to fetch a cacheline
per-block. The compiler could then use some sort of unpacking code to create
the right xmm values.

Surprisingly, now the compiler is able to do even better with SSE 4.1.
pmovzxwd pulls a word from memory and expands it out to an 128b value.

Before:
182:   49 c1 e0 05             shl    r8,0x5
186:   66 43 0f 6f 94 30 40    movdqa xmm2,XMMWORD PTR [r8+r14*1+0x340]

After:
182:   49 c1 e0 04             shl    r8,0x4
186:   66 43 0f 38 33 94 30    pmovzxwd xmm2,QWORD PTR [r8+r14*1+0x340]

+++ src/mesa/main/texcompress_astc.cpp
@@ -650,7 +650,7 @@ struct Block
    uint8_t infill_weights[2][216]; /* large enough for 6x6x6 */
 
    /* Calculated by decode_colour_endpoints(); */
-   uint32_t endpoints_decoded[4][2][4] __attribute__((aligned(64)));
+   uint16_t endpoints_decoded[4][2][4] __attribute__((aligned(64)));
 
    void calculate_from_weights();
    void calculate_remaining_bits();
